Separate `Linq` Tests and add additional exceptions from unexpected arguments.
